fixed some of the MSVC warnings

+++ src/candy/core/Solver.h
@@ -681,7 +681,7 @@ bool Solver<PickBranchLitT>::isIncremental() {
  ***/
 template<class PickBranchLitT>
 Var Solver<PickBranchLitT>::newVar(bool sign, bool dvar) {
-    int v = nVars();
+    int v = static_cast<int>(nVars());
     watches.init(mkLit(v, false));
     watches.init(mkLit(v, true));
     watchesBin.init(mkLit(v, false));
